# [PERF] cuda.parallel: Cache intermediate results to improve performance of `cudax.reduce_into` 