## Description

This PR uses caching on the Python side to improve the performance of `cuda.experimental.reduce_into`. Specifically:

- cache `_Reduce` objects. The cache key used here is the `dtype` of the input arrays rather than the arrays themselves. I _think_ this is safe to do, and longer term, I'd like to avoid passing the concrete arrays to the `_Reduce` constructor. 

- cache the result of the utility function `_type_to_info`.

Before this PR:

```python
In [4]: d_in = cuda.device_array(1, "int64")

In [5]: d_out = cuda.device_array(1, "int64")

In [6]: h_init = np.asarray([0], "int64")

In [7]: def op(x, y): return x + y

In [8]: %timeit cudax.reduce_into(d_in, d_out, op, h_init)
781 ms ± 42.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [9]: %timeit cudax._type_to_info(np.int32)
71.8 μs ± 111 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

After this PR:

```python
In [8]: %timeit cudax.reduce_into(d_in, d_out, op, h_init)
The slowest run took 17.54 times longer than the fastest. This could mean that an intermediate result is being cached.
910 ns ± 1.36 μs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [9]: %timeit cudax._type_to_info(np.int32)
64.8 ns ± 0.0356 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
```



## Checklist
<!-- TODO: - [ ] I am familiar with the [Contributing Guidelines](). -->
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.