@shwina 2024-12-02
Looks like the build docs CI job is using Python=3.7 which is missing the `functools.cache` function. Looking into it.